[sheets-] simplify column autoname during newSheet
