+++ visidata/sheets.py
@@ -8,6 +8,7 @@ from visidata import (options, Column, namedlist, SettableColumn, AttrDict, Disp
 TypedExceptionWrapper, BaseSheet, UNLOADED, wrapply,
 clipdraw, clipdraw_chunks, ColorAttr, update_attr, colors, undoAttrFunc, vlen, dispwidth)
 import visidata
+from visidata.utils import colname_letters
 
 
 vd.activePane = 1   # pane numbering starts at 1; pane 0 means active pane
@@ -1189,14 +1190,9 @@ def preloadHook(sheet):
 
 @VisiData.api
 def newSheet(vd, name, ncols, **kwargs):
-    #use temporary column names until we have the sheet object that can assign names and keep track of them
-    rename_cols = [SettableColumn(width=vd.options.default_width, name=f'name{i}') for i in range(ncols)]
-    vs = Sheet(name, columns=rename_cols, **kwargs)
-    # For TableSheet, we cannot change colnames using setName() on vs.columns[i],
-    # because later in reload(), TableSheet.resetCols() recopies the names from the original
-    # column objects. So we have to run setName() on the original column objects.
-    for c in rename_cols:
-        c.setName(vs.incremented_colname())
+    cols = [SettableColumn(width=vd.options.default_width, name=f'{colname_letters(i+1)}') for i in range(ncols)]
+    vs = Sheet(name, columns=cols, **kwargs)
+    vs.colname_ctr = ncols
     return vs
 
 @BaseSheet.api
